Fix erreurs de compilation
Il est possible qu'au futur on devra encore ajouter '-march=pentium', supprimer '-gstabs' et/ou changer l'utilisation des `__attribute__`.

+++ user/tests/lib/sysapi.h
@@ -128,7 +128,7 @@ void sys_info(void);
 typedef __SIZE_TYPE__ size_t;
 
 int strcmp(const char *str1, const char *str2);
-unsigned long strlen(const char *s);
+size_t strlen(const char *s);
 char *strncpy(char *dst, const char *src, unsigned n);
 void *memset(void *dst, int c, size_t n);
